next

+++ MS/microservice-produits/src/main/java/com/mcommerce/microservice_produits/dao/ProductDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductDao extends JpaRepository<Product, Integer> {}
+public interface ProductDao extends JpaRepository<Product,Long> {}
 
 //Le DAO gère l’accès aux données via JPA (findAll, findById, save...).
 //Ce DAO permet d’effectuer automatiquement toutes les opérations CRUD sans écrire une seule requête SQL.
